services/locationService/requirements.txt
pyzmq==27.1.0
python-dotenv==1.2.2
geopy==2.4.1